# petclinic-java-gradle-react-tailwind-ts/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Petclinic API
  version: 1.0.0
  description: REST API for the Petclinic application
servers:
  - url: '{openapi.petclinic.base-path:}'
    variables:
      openapi.petclinic.base-path:
        default: ""
...
paths:
...
  /api/vets:
    get:
      tags:
        - Vet
      operationId: getAllVets
      x-spring-paginated: true
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Vet
      operationId: createVet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VetDTO'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetCreateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/vets/{id}:
    get:
      tags:
        - Vet
      operationId: getVet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetGetResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags:
        - Vet
      operationId: updateVet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VetDTO'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VetUpdateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Vet
      operationId: deleteVet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/types:
    get:
      tags:
        - Type
      operationId: getAllTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Type
      operationId: createType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeDTO'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeCreateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/types/{id}:
    get:
      tags:
        - Type
      operationId: getType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeGetResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags:
        - Type
      operationId: updateType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeDTO'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeUpdateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Type
      operationId: deleteType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/visits:
    get:
      tags:
        - Visit
      operationId: getAllVisits
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitListResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
        - Visit
      operationId: createVisit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitDTO'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitCreateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/visits/{id}:
    get:
      tags:
        - Visit
      operationId: getVisit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitGetResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    put:
      tags:
        - Visit
      operationId: updateVisit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitDTO'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitUpdateResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags:
        - Visit
      operationId: deleteVisit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        error:
          type: string

    # ---------------------------------------------------------------------------
    # Owner
    # ---------------------------------------------------------------------------
...
    VetDTO:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        vetSpecialtySpecialties:
          type: array
          items:
            type: integer
      required:
        - firstName
        - lastName
        - createdAt
        - updatedAt
    VetListData:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VetDTO'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    VetListResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          $ref: '#/components/schemas/VetListData'
    VetGetResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          $ref: '#/components/schemas/VetDTO'
    VetCreateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer
    VetUpdateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer

    # ---------------------------------------------------------------------------
    # Type
    # ---------------------------------------------------------------------------
    TypeDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - name
        - createdAt
        - updatedAt
...
    TypeListResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/TypeDTO'
    TypeGetResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          $ref: '#/components/schemas/TypeDTO'
    TypeCreateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer
    TypeUpdateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer

    # ---------------------------------------------------------------------------
    # Visit
    # ---------------------------------------------------------------------------
    VisitDTO:
      type: object
      properties:
        id:
          type: integer
        visitDate:
          type: string
          format: date
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        pet:
          type: integer
      required:
        - visitDate
        - description
        - createdAt
        - updatedAt
        - pet
...
    VisitListResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/VisitDTO'
    VisitGetResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          $ref: '#/components/schemas/VisitDTO'
    VisitCreateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer
    VisitUpdateResponse:
      x-implements:
        - dev.pollito.petclinic_java_gradle_react_tailwind_ts.util.metadata.ResponseMetadata
      type: object
      properties:
        instance:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        trace:
          type: string
        data:
          type: integer

    # ---------------------------------------------------------------------------
    # Specialty
    # ---------------------------------------------------------------------------
...